Switch int to long in sales

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = { 10, 2, 333, 44 };
+        long[] sales = { 10, 2, 333, 44 };
         SalesManager manager = new SalesManager(sales);
         System.out.println("Макс: " + manager.max());
     }
